secrets/requirements.txt
pynacl==1.4
argparse==1.4